+++ packages/plugin-rsc/examples/basic/src/routes/serialization/action.tsx
@@ -1,0 +1,6 @@
+'use server'
+
+export async function testSerializationAction() {
+  console.log('[test-serialization-action]')
+  return 'ok'
+}