+++ packages/plugin-rsc/examples/basic/test-dep/client-in-server/server.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import { TestClient } from './client.js'
+
+export async function TestClientInServerDep() {
+  return React.createElement(TestClient)
+}