+++ packages/plugin-rsc/examples/basic/test-dep/client-in-server2/server.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import { TestContextProvider } from './client.js'
+
+export function TestContextProviderInServer(props) {
+  return React.createElement(
+    TestContextProvider,
+    { value: props.value },
+    props.children,
+  )
+}