+++ playground/react-emotion/__tests__/react.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test } from 'vitest'
-import { editFile, getColor, isServe, page, untilUpdated } from '~utils'
+import { editFile, getColor, isServe, page } from '~utils'
 
 test('should render', async () => {
   expect(await page.textContent('h1')).toMatch(
@@ -17,13 +17,13 @@ test.runIf(isServe)('should hmr', async () => {
   editFile('App.jsx', (code) =>
     code.replace('Vite + React + @emotion/react', 'Updated'),
   )
-  await untilUpdated(() => page.textContent('h1'), 'Hello Updated')
+  await expect.poll(() => page.textContent('h1')).toMatch('Hello Updated')
 
   editFile('Counter.jsx', (code) =>
     code.replace('color: #646cff;', 'color: #d26ac2;'),
   )
 
-  await untilUpdated(() => getColor('code'), '#d26ac2')
+  await expect.poll(() => getColor('code')).toMatch('#d26ac2')
 
   // preserve state
   expect(await page.textContent('button')).toMatch('count is: 1')
@@ -47,7 +47,7 @@ test('should update button style', async () => {
       code.replace('border: 2px solid #000', 'border: 4px solid red'),
     )
 
-    await untilUpdated(getButtonBorderStyle, '4px solid rgb(255, 0, 0)')
+    await expect.poll(getButtonBorderStyle).toMatch('4px solid rgb(255, 0, 0)')
 
     // preserve state
     expect(await page.textContent('button')).toMatch('count is: 1')